- img_preproc   * align/alignimg : image crop parameters are read from opts struct   * align_face : now supports multiple faces detected in one image   * facedet/runfacedet.m : use precomputed vj file if exist   * init.m : changing the image crop parameters to get tight face image of size 152 x 152
- cfw/AlignDataset.m : align all images from dataset
- cfw/ReadBbFile.m : read bounding box file (used as gt for the face detection process)

+++ img_preproc/facedet/runfacedet.m
@@ -8,13 +8,16 @@ else
     detpath=[imgpath '.vj'];
 end
 
-imwrite(I,pgmpath);
+if ~exist(detpath, 'file')
+    % run this only if the detections file doesn;t exist from previous run
+    imwrite(I,pgmpath);
 
-root=[fileparts(which(mfilename)) '/OpenCV_ViolaJones'];
-system(sprintf('%s/Release/OpenCV_ViolaJones %s/haarcascade_frontalface_alt.xml %s %s',root,root,pgmpath,detpath));
+    root=fullfile(fileparts(which(mfilename)), 'OpenCV_ViolaJones');
+    system(sprintf('%s/Release/OpenCV_ViolaJones %s/haarcascade_frontalface_alt.xml "%s" "%s"',root,root,pgmpath,detpath));
+    delete(pgmpath);
+end
 
 DETS=readfacedets(detpath);
-delete(pgmpath);
 if nargin<2
     delete(detpath);
 end
